% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ns.r
\docType{data}
\name{example.mat.2D}
\alias{example.mat.2D}
\title{2-dimensional matern example data}
\format{A matrix.}
\usage{
example.mat.2D
}
\description{
Simulated data, with children points generated from a Poisson(10)
}
\keyword{datasets}
